+++ source/core/hist/histogram_manager/HistogramManagerMTFFGridSurface.cpp
@@ -61,10 +61,10 @@ std::unique_ptr<ICompositeDistanceHistogram> HistogramManagerMTFFGridSurface<var
 
     auto& data_a = *this->data_a_ptr;
     auto& data_w = *this->data_w_ptr;
-    int data_a_size = (int) data_a.size();
-    int data_w_size = (int) data_w.size();
-    int data_x_i_size = (int) data_x_i.size();
-    int data_x_s_size = (int) data_x_s.size();
+    int data_a_size = data_a.size();
+    int data_w_size = data_w.size();
+    int data_x_i_size = data_x_i.size();
+    int data_x_s_size = data_x_s.size();
     int bin_count = hist::detail::required_bin_count<variable_bin_width>(data_a, data_w, data_x_i, data_x_s);
 
     //########################//
